Conflicts:
	test/TestHelper.sol

+++ mocks/wells/MockReserveWell.sol
@@ -1,4 +1,6 @@
-// SPDX-License-Identifier: MIT
+/**
+ * SPDX-License-Identifier: MIT
+ */
 
 pragma solidity ^0.8.17;
 
@@ -8,22 +10,21 @@ import {IPump} from "src/interfaces/pumps/IPump.sol";
  * @notice Mock Well that allows setting of reserves.
  */
 contract MockReserveWell {
-
-    uint[] reserves;
+    uint256[] reserves;
 
     constructor() {
-        reserves = new uint[](2);
+        reserves = new uint256[](2);
     }
 
-    function setReserves(uint[] memory _reserves) public {
+    function setReserves(uint256[] memory _reserves) public {
         reserves = _reserves;
     }
 
-    function getReserves() external view returns (uint[] memory _reserves) {
+    function getReserves() external view returns (uint256[] memory _reserves) {
         _reserves = reserves;
     }
 
-    function update(address pump, uint[] calldata _reserves, bytes calldata data) external {
+    function update(address pump, uint256[] calldata _reserves, bytes calldata data) external {
         IPump(pump).update(reserves, data);
         setReserves(_reserves);
     }
