+++ script/deploy/MockPump.s.sol
@@ -8,7 +8,7 @@ import {MockPump} from "mocks/pumps/MockPump.sol";
 // Mockpump does not provide utility and is solely used for example.
 contract DeployMockPump is Script {
     function run() external {
-        uint deployerPrivateKey = vm.envUint("PRIVATE_KEY");
+        uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
         vm.startBroadcast(deployerPrivateKey);
 
         MockPump mockPump = new MockPump();
