+++ test/Well.Sync.t.sol
@@ -4,15 +4,15 @@ pragma solidity ^0.8.17;
 import {MockToken, TestHelper, Balances} from "test/TestHelper.sol";
 
 contract WellSyncTest is TestHelper {
-    event Sync(uint[] reserves);
+    event Sync(uint256[] reserves);
 
     function setUp() public {
         setupWell(2);
 
         // Let `user` burn
         vm.startPrank(address(well));
-        MockToken(address(tokens[0])).approve(address(user), type(uint).max);
-        MockToken(address(tokens[1])).approve(address(user), type(uint).max);
+        MockToken(address(tokens[0])).approve(address(user), type(uint256).max);
+        MockToken(address(tokens[1])).approve(address(user), type(uint256).max);
         vm.stopPrank();
     }
 
@@ -26,7 +26,7 @@ contract WellSyncTest is TestHelper {
         MockToken(address(tokens[0])).burnFrom(address(well), 1e18);
         MockToken(address(tokens[1])).burnFrom(address(well), 1e18);
 
-        uint[] memory expectedReserves = new uint[](2);
+        uint256[] memory expectedReserves = new uint256[](2);
         expectedReserves[0] = 999e18;
         expectedReserves[1] = 999e18;
 
@@ -35,7 +35,7 @@ contract WellSyncTest is TestHelper {
 
         well.sync();
 
-        uint[] memory reserves = well.getReserves();
+        uint256[] memory reserves = well.getReserves();
         assertEq(reserves[0], expectedReserves[0], "Reserve 0 should be 1e18");
         assertEq(reserves[1], expectedReserves[1], "Reserve 1 should be 1e18");
     }
@@ -44,7 +44,7 @@ contract WellSyncTest is TestHelper {
         MockToken(address(tokens[0])).mint(address(well), 1e18);
         MockToken(address(tokens[1])).mint(address(well), 1e18);
 
-        uint[] memory expectedReserves = new uint[](2);
+        uint256[] memory expectedReserves = new uint256[](2);
         expectedReserves[0] = 1001e18;
         expectedReserves[1] = 1001e18;
 
@@ -53,13 +53,13 @@ contract WellSyncTest is TestHelper {
 
         well.sync();
 
-        uint[] memory reserves = well.getReserves();
+        uint256[] memory reserves = well.getReserves();
         assertEq(reserves[0], expectedReserves[0], "Reserve 0 should be Balance 0");
         assertEq(reserves[1], expectedReserves[1], "Reserve 1 should be Balance 1");
     }
 
     function testFuzz_sync(uint128[2] calldata mintAmount, uint128[2] calldata burnAmount) public prank(user) {
-        uint temp = bound(mintAmount[0], 0, type(uint128).max - tokens[0].balanceOf(address(well)) - 100);
+        uint256 temp = bound(mintAmount[0], 0, type(uint128).max - tokens[0].balanceOf(address(well)) - 100);
         MockToken(address(tokens[0])).mint(address(well), temp);
         temp = bound(mintAmount[1], 0, type(uint128).max - tokens[1].balanceOf(address(well)) - 100);
         MockToken(address(tokens[1])).mint(address(well), temp);
@@ -68,7 +68,7 @@ contract WellSyncTest is TestHelper {
         temp = bound(burnAmount[1], 0, tokens[1].balanceOf(address(well)));
         MockToken(address(tokens[1])).burnFrom(address(well), temp);
 
-        uint[] memory balances = new uint[](2);
+        uint256[] memory balances = new uint256[](2);
         balances[0] = tokens[0].balanceOf(address(well));
         balances[1] = tokens[1].balanceOf(address(well));
 
@@ -77,7 +77,7 @@ contract WellSyncTest is TestHelper {
 
         well.sync();
 
-        uint[] memory reserves = well.getReserves();
+        uint256[] memory reserves = well.getReserves();
         assertEq(reserves[0], balances[0], "Reserve 0 should be Balance 0");
         assertEq(reserves[1], balances[1], "Reserve 1 should be Balance 1");
     }
